ml6team_mlsd/requirements.txt
streamlit==1.20.0
streamlit-drawable-canvas==0.9.0
diffusers==0.15.0
xformers==0.0.16
transformers==4.28.0
torchvision==0.14.1
git+https://github.com/huggingface/accelerate.git
opencv-python-headless==4.7.0.72
scipy==1.10.0
python-docx
extra-streamlit-components==0.1.56
triton
altair==4.1.0
